requirements.txt
# Django
asgiref==3.3.4
Django==3.2.4

# enviroments var
django-environ==0.4.5

# PostgreSQL
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1

# Filters 
django-filter==2.4.0

# django rest framework 
djangorestframework==3.12.4

# Http Requests
certifi==2021.5.30
idna==2.10
requests==2.25.1
urllib3==1.26.5